requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-socketio==5.3.6
eventlet==0.35.1
pyyaml==6.0.2
scikit-learn==1.5.2
numpy>=1.26.0
python-dotenv==1.0.1
gunicorn==22.0.0
pytest==8.3.2
